chore: #221 타입 변경

+++ perflow-backend/src/main/java/com/touchdown/perflowbackend/payment/command/application/dto/InsuranceRateRequestDTO.java
@@ -11,12 +11,12 @@ public class InsuranceRateRequestDTO {
 
     private final Long insuranceId;
 
-    private final Long nationalPensionRate;
+    private final double nationalPensionRate;
 
-    private final Long healthInsuranceRate;
+    private final double healthInsuranceRate;
 
-    private final Long hireInsuranceRate;
+    private final double hireInsuranceRate;
 
-    private final Long longTermCareInsuranceRate;
+    private final double longTermCareInsuranceRate;
 
 }
